python-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
yt-dlp==2023.11.16
openai-whisper==20231117
pydub==0.25.1
supabase==2.0.2
python-dotenv==1.0.0
